Trivial: rm_Server
The "room" for the server can run at any speed. For example it is set to
100 frames per second ( our client runs at 30 ). This should allow it to
process network queues faster than it needs to so it won't cause a
bottleneck of lag. To keep things in sync, the world broadcast is
manually calculated ( but can be automatically calculated if you want )
to 100ms. So all clients will be 100ms behind, together. Interpolation!

THIS CRAP FINALLY WORKS TODAY WAS VERY FRUSTRATING. THIS IS NOT A
PROFESSIONAL COMMIT. I DONT CARE. IT WORKS.