Feature: obj_Server
* Code Cleanup: Now using the built in alarms feature for better code
organization. I was able to split up the scripts for readability.
* The server now requests its external IP / Port / Name upon starting
and connects to a List Server. This means that Servers can be hosted
from multiple locations and players will still connect to an initial
centralized place. This also means that more than one server can be
running at the same time.